Fetch data in UserReservationsPage container
Previously reservations, resources and units were fetched in
ReservationsList container. In the future UserReservationsPage will
have more than one reservations list, at least for admins, so it's good
to do the data fetching only once in one single place.
Refs #248.
