+++ app/containers/UserReservationsPage.js
@@ -1,9 +1,20 @@
-import React, { Component } from 'react';
+import React, { Component, PropTypes } from 'react';
 import DocumentTitle from 'react-document-title';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
 
+import { fetchReservations } from 'actions/reservationActions';
+import { fetchResources } from 'actions/resourceActions';
+import { fetchUnits } from 'actions/unitActions';
 import ReservationsList from 'containers/ReservationsList';
 
-class UserReservationsPage extends Component {
+export class UnconnectedUserReservationsPage extends Component {
+  componentDidMount() {
+    this.props.actions.fetchResources();
+    this.props.actions.fetchUnits();
+    this.props.actions.fetchReservations({ isOwn: true });
+  }
+
   render() {
     return (
       <DocumentTitle title="Omat varaukset - Varaamo">
@@ -16,6 +27,18 @@ class UserReservationsPage extends Component {
   }
 }
 
-UserReservationsPage.propTypes = {};
+UnconnectedUserReservationsPage.propTypes = {
+  actions: PropTypes.object.isRequired,
+};
 
-export default UserReservationsPage;
+function mapDispatchToProps(dispatch) {
+  const actionCreators = {
+    fetchReservations,
+    fetchResources,
+    fetchUnits,
+  };
+
+  return { actions: bindActionCreators(actionCreators, dispatch) };
+}
+
+export default connect(null, mapDispatchToProps)(UnconnectedUserReservationsPage);
